refactor: remove `manager.rs`
For now, the process manager is integrated to the kernel.
